Add signals
