--- Games/MysticHeroes2/src/Engine/IObserver.h
@@ -1,9 +1,0 @@
-#pragma once
-
-class IObserver
-{
-public:
-    virtual void OnNext(const void* dataPointer) { }
-
-    virtual ~IObserver() = default;
-};
