--- Games/MysticHeroes2/src/Engine/Subscription.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "IObserver.h"
-
-class Observable;
-
-class Subscription
-{
-public:
-    Subscription(Observable* signal, IObserver* observer);
-
-    ~Subscription();
-
-private:
-    Observable* _signal;
-    IObserver* _observer;
-};
