+++ Games/Sandbox/imgui.ini
@@ -1,0 +1,10 @@
+[Window][Debug##Default]
+Pos=60,60
+Size=400,400
+Collapsed=0
+
+[Window][Hello, world!]
+Pos=60,60
+Size=339,111
+Collapsed=0
+
